requirements.txt
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
drf-spectacular==0.28.0
python-dotenv==1.1.1